First test passing on iPhone safari as well.

+++ src/test/groovy/impl/Constants.groovy
@@ -4,8 +4,8 @@ package impl
 class Constants {
 
     static final String ENVIRONMENT_URL = "https://www.coolbet.com"
-    static final int TIMEOUT10S = 10
-    static final int TIMEOUT2S= 2
-    static final int TIMEOUT20S= 20
-    static final int TIMEOUT30S= 30
+    static final long TIMEOUT10S = 10
+    static final long TIMEOUT2S = 2
+    static final long TIMEOUT20S = 20
+    static final long TIMEOUT30S = 30
 }
