+++ src/test/groovy/impl/betslip/pageobjects/BetSlip.groovy
@@ -2,7 +2,6 @@ package impl.betslip.pageobjects
 
 import io.cify.framework.PageObjects
 import io.cify.framework.core.Device
-import org.openqa.selenium.By
 import org.openqa.selenium.WebElement
 import org.openqa.selenium.support.FindBy
 
@@ -33,7 +32,7 @@ class BetSlip extends PageObjects{
         super(device)
     }
 
-    BetSlip(Device device, int timeout){
+    BetSlip(Device device, long timeout) {
         super(device, timeout)
     }
 }
